requirements.txt
accelerate
aniso8601
attrs
bitsandbytes
blinker
certifi
charset-normalizer
clams-python==1.1.3
click
deepdiff
ffmpeg-python
filelock
Flask
Flask-RESTful
fsspec
future
gunicorn
huggingface-hub
lxml
MarkupSafe
mmif-python==1.0.10
mpmath
numpy
opencv-python
ordered-set
packaging
pillow
platformdirs
pydantic
pytz
PyYAML
referencing
regex
safetensors
sympy
tokenizers
torch
tqdm
transformers
